+++ config/application.rb
@@ -30,7 +30,7 @@ module Events
     # Autoload paths
     # You typically add directories to autoload_paths so that classes and modules are loaded
     # automatically as they are referenced. This makes development more convenient
-    # config.autoload_paths << "#{config.root}/app/workers"
+    # config.autoload_paths += %W(#{config.root}/app/models/concerns)
 
     # Eager load paths
     # You typically add directories to eager_load_paths to ensure that all necessary classes
