+++ spec/factories/event_factory.rb
