+++ spec/utlilities/doi_utilities_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# RSpec.describe(DoiUtilities) do
+#   describe "#normalize_doi" do
+#     it "returns nil if the doi is not found"
+#     end
+#   end
+# end
